🤹‍♀️ Move test app source files
