+++ projects/test-app/src/favicon.ico
