+++ projects/test-app/src/pharkas-test.component.ts
